+++ src/test/java/com/frynd/debouncer/accumulator/impl/MapAccumulatorTest.java
@@ -1,7 +1,7 @@
 package com.frynd.debouncer.accumulator.impl;
 
 import com.frynd.debouncer.accumulator.Accumulator;
-import com.frynd.debouncer.accumulator.Drainers;
+import com.frynd.debouncer.drainer.Drainers;
 import com.frynd.debouncer.accumulator.Record;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
